Move installWrangler to packages
