+++ packages/create-cloudflare/src/helpers/__tests__/command.test.ts
@@ -2,7 +2,7 @@ import { existsSync } from "fs";
 import { spawn } from "cross-spawn";
 import { afterEach, beforeEach, describe, expect, test, vi } from "vitest";
 import whichPMRuns from "which-pm-runs";
-import { installWrangler, runCommand } from "../command";
+import { runCommand } from "../command";
 import type { ChildProcess } from "child_process";
 
 // We can change how the mock spawn works by setting these variables
@@ -44,18 +44,8 @@ describe("Command Helpers", () => {
 		});
 
 		vi.mocked(whichPMRuns).mockReturnValue({ name: "npm", version: "8.3.1" });
-
 		vi.mocked(existsSync).mockImplementation(() => false);
 	});
-
-	const expectSilentSpawnWith = (cmd: string) => {
-		const [command, ...args] = cmd.split(" ");
-
-		expect(spawn).toHaveBeenCalledWith(command, args, {
-			stdio: "pipe",
-			env: process.env,
-		});
-	};
 
 	test("runCommand", async () => {
 		await runCommand(["ls", "-l"]);
@@ -63,11 +53,5 @@ describe("Command Helpers", () => {
 			stdio: "inherit",
 			env: process.env,
 		});
-	});
-
-	test("installWrangler", async () => {
-		await installWrangler();
-
-		expectSilentSpawnWith("npm install --save-dev wrangler");
 	});
 });
